% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVa.R
\name{calcVa}
\alias{calcVa}
\title{Calculate Va}
\usage{
calcVa(afs, betas)
}
\arguments{
\item{afs}{vector of mean allele frequencies (mafs or afs are ok)}

\item{betas}{vector of effect sizes}
}
\description{
This function calculates additive genetic variation (Va) from a set of loci effect sizes and mean allele frequencies
}
