% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_q_values.R
\name{get_q_values}
\alias{get_q_values}
\title{Calculate qvalues for a table of values}
\usage{
get_q_values(ptable)
}
\arguments{
\item{ptable}{table of values}
}
\description{
Calculates qvalues for a table of values
}
