% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaf.R
\name{getMaf}
\alias{getMaf}
\title{Calculate minor allele frequency}
\usage{
getMaf(af)
}
\arguments{
\item{af}{an allele frequency}
}
\description{
This function takes an allele frequency (range 0 to 1) and calculates the minor allele frequency (0 to 0.5)
}
