% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_k_complete.R
\name{make_k_complete}
\alias{make_k_complete}
\title{Kinship matrix function for complete data}
\usage{
make_k_complete(myG)
}
\arguments{
\item{myG}{matrix where the rows are individuals/populations and the columns are loci and the values are the allele frequency (not the # of copies present in an individual!!!).}
}
\description{
This function makes a kinship matrix using the cov function.
}
