% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_k_E.R
\name{make_k_E}
\alias{make_k_E}
\title{Kinship matrix function for complete data using the estimated variance across all loci}
\usage{
make_k_E(myG)
}
\arguments{
\item{myG}{matrix where the rows are individuals/populations and the columns are loci and the values are the allele frequency (not the # of copies present in an individual!!!).}
}
\description{
This function makes a kinship matrix using the cov function. Unlike make_k_complete, it standardizes by the estimated genic variance across all loci, not each locus individually
}
