% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGemmaOutput.R
\name{processGemmaOutput}
\alias{processGemmaOutput}
\title{Process Gemma Output}
\usage{
processGemmaOutput(x)
}
\arguments{
\item{x}{the path to the gemma file}
}
\description{
This function takes an output file from gemma and turns it into a useable data.frame
}
