% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var0.R
\name{var0}
\alias{var0}
\title{Calculate variance where the mean is set to zero}
\usage{
var0(x)
}
\arguments{
\item{x}{a string of vectors}
}
\description{
This function takes a string of numbers and calculates the variance of these numbers, assuming that the mean is 0.
}
