% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQpc.R
\name{calcQpc}
\alias{calcQpc}
\title{Calculate Qpc}
\usage{
calcQpc(myZ, myU, myLambdas, myPCcutoff = 0.3)
}
\arguments{
\item{myZ}{vector of traits}

\item{myU}{matrix of eigenvectors of the kinship matrix (each column is an eigenvector)}

\item{myLambdas}{vector of eigenvalues of the kinship matrix}

\item{myPCcutoff}{a value that determines how many PCs you want to look at. For example, 0.5 would mean that you would look at the first set of PCs that explain 0.5 of the variation. The default here is 0.3}
}
\description{
This function calculates Qpc given data about the relatedness matrix, and a set of trait values
}
\examples{
calcQpc()
}
